Scripts/ingest/pyproject.toml
[project]
name = "voice-note-ingest"
version = "1.0.0"
description = "Production voice note ingestion pipeline"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "watchdog>=3.0.0",
]

[project.scripts]
ingest = "ingest.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"